Move FSO combinations definition to OzoneContract
