+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/fs/ozone/contract/ITestOzoneContractGetFileStatus.java
@@ -78,6 +78,6 @@ public class ITestOzoneContractGetFileStatus
 
   @Parameterized.Parameters
   public static Collection<Boolean> data() {
-    return ITestOzoneContractUtils.getFsoCombinations();
+    return OzoneContract.getFsoCombinations();
   }
 }
